pyproject.toml
[tool.poetry]
name = "sgd-dstc8-data-model"
version = "0.1.9"
description = "Model Classes for the SGD DSTC8 dataset"
authors = ["Adib Mosharrof <adib.mosharrof@gmail.com>"]
repository = "https://github.com/adibMosharrof/sgd_dstc8_data_model"
readme = "README.md"
packages = [{include = "sgd_dstc8_data_model"}]
keywords = ["sgd", "dstc8", "data", "model", "classes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3.9",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[tool.poetry.dependencies]
python = "^3.10"
dataclasses-json = "^0.5.7"
pyyaml = "^6.0"
class-registry = "^2.1.2"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
pytest = "^7.2.2"

[tool.poetry.plugins."dstc_special_tokens"]
dstc_special_tokens = "dstc_constants:DstcSpecialTokens"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
